backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg[binary]>=3.1.0
Pillow>=10.2.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
python-decouple==3.8
django-filter==23.5
django-extensions==3.2.3
drf-spectacular==0.27.2
django-storages==1.14.4
boto3>=1.34.0
python-jose[cryptography]>=3.3.0
python-jwt>=4.0.0
cryptography>=41.0.0
PyJWT>=2.8.0
requests>=2.31.0
